Enable Speculative Hold for RSFT, RCTL by default.
