+++ tests/tap_hold_configurations/speculative_hold/all_mods/config.h
@@ -20,3 +20,4 @@
 #include "test_common.h"
 
 #define SPECULATIVE_HOLD
+#define DUMMY_MOD_NEUTRALIZER_KEYCODE KC_F24
