+++ tests/tap_hold_configurations/speculative_hold/default/test_tap_hold.cpp
@@ -27,28 +27,13 @@ using testing::InSequence;
 
 namespace {
 
-// Gets the unpacked 8-bit mods corresponding to a given mod-tap keycode.
-uint8_t unpack_mod_tap_mods(uint16_t keycode) {
-    const uint8_t mods5 = QK_MOD_TAP_GET_MODS(keycode);
-    return (mods5 & 0x10) != 0 ? (mods5 << 4) : mods5;
-}
-
-bool get_speculative_hold_all_keys(uint16_t keycode, keyrecord_t *record) {
-    return true; // Enable Speculative Hold for all mod-tap keys.
-}
-
 bool process_record_user_default(uint16_t keycode, keyrecord_t *record) {
     return true;
 }
 
-// Indirection so that get_speculative_hold() and process_record_user() can be
+// Indirection so that process_record_user() can be
 // replaced with other functions in the test cases below.
-std::function<bool(uint16_t, keyrecord_t *)> get_speculative_hold_fun = get_speculative_hold_all_keys;
-std::function<bool(uint16_t, keyrecord_t *)> process_record_user_fun  = process_record_user_default;
-
-extern "C" bool get_speculative_hold(uint16_t keycode, keyrecord_t *record) {
-    return get_speculative_hold_fun(keycode, record);
-}
+std::function<bool(uint16_t, keyrecord_t *)> process_record_user_fun = process_record_user_default;
 
 extern "C" bool process_record_user(uint16_t keycode, keyrecord_t *record) {
     return process_record_user_fun(keycode, record);
@@ -57,10 +42,29 @@ extern "C" bool process_record_user(uint16_t keycode, keyrecord_t *record) {
 class SpeculativeHoldDefault : public TestFixture {
    public:
     void SetUp() override {
-        get_speculative_hold_fun = get_speculative_hold_all_keys;
-        process_record_user_fun  = process_record_user_default;
+        process_record_user_fun = process_record_user_default;
     }
 };
+
+TEST_F(SpeculativeHoldDefault, get_speculative_hold) {
+    keyrecord_t record = {};
+
+    // With the default definition of get_speculative_hold(), Speculative Hold
+    // is enabled for Ctrl and Shift.
+    EXPECT_TRUE(get_speculative_hold(MT(MOD_LCTL, KC_NO), &record));
+    EXPECT_TRUE(get_speculative_hold(MT(MOD_LSFT, KC_NO), &record));
+    EXPECT_TRUE(get_speculative_hold(MT(MOD_LCTL | MOD_LSFT, KC_NO), &record));
+    EXPECT_TRUE(get_speculative_hold(MT(MOD_RCTL, KC_NO), &record));
+    EXPECT_TRUE(get_speculative_hold(MT(MOD_RSFT, KC_NO), &record));
+    EXPECT_TRUE(get_speculative_hold(MT(MOD_RCTL | MOD_RSFT, KC_NO), &record));
+
+    EXPECT_FALSE(get_speculative_hold(MT(MOD_LALT, KC_NO), &record));
+    EXPECT_FALSE(get_speculative_hold(MT(MOD_LGUI, KC_NO), &record));
+    EXPECT_FALSE(get_speculative_hold(MT(MOD_RALT, KC_NO), &record));
+    EXPECT_FALSE(get_speculative_hold(MT(MOD_RGUI, KC_NO), &record));
+    EXPECT_FALSE(get_speculative_hold(MT(MOD_MEH, KC_NO), &record));
+    EXPECT_FALSE(get_speculative_hold(MT(MOD_HYPR, KC_NO), &record));
+}
 
 TEST_F(SpeculativeHoldDefault, tap_mod_tap) {
     TestDriver driver;
@@ -100,232 +104,6 @@ TEST_F(SpeculativeHoldDefault, tap_mod_tap) {
 
     // Idle for tapping term of mod tap hold key.
     idle_for(TAPPING_TERM - 10);
-    VERIFY_AND_CLEAR(driver);
-}
-
-TEST_F(SpeculativeHoldDefault, tap_mod_tap_neutralized) {
-    TestDriver driver;
-    InSequence s;
-    auto       mod_tap_key = KeymapKey(0, 1, 0, GUI_T(KC_P));
-
-    set_keymap({mod_tap_key});
-
-    // Press mod-tap-hold key. Mod is held speculatively.
-    EXPECT_REPORT(driver, (KC_LGUI));
-    mod_tap_key.press();
-    idle_for(10);
-    VERIFY_AND_CLEAR(driver);
-
-    // Release mod-tap-hold key. Speculative mod is neutralized and canceled.
-    EXPECT_REPORT(driver, (KC_LGUI, DUMMY_MOD_NEUTRALIZER_KEYCODE));
-    EXPECT_REPORT(driver, (KC_LGUI));
-    EXPECT_EMPTY_REPORT(driver);
-    EXPECT_REPORT(driver, (KC_P));
-    EXPECT_EMPTY_REPORT(driver);
-    mod_tap_key.release();
-    run_one_scan_loop();
-    VERIFY_AND_CLEAR(driver);
-
-    // Idle for tapping term of mod tap hold key.
-    idle_for(TAPPING_TERM - 10);
-    VERIFY_AND_CLEAR(driver);
-}
-
-TEST_F(SpeculativeHoldDefault, hold_two_mod_taps) {
-    TestDriver driver;
-    InSequence s;
-    auto       mod_tap_key1 = KeymapKey(0, 1, 0, LCTL_T(KC_A));
-    auto       mod_tap_key2 = KeymapKey(0, 2, 0, RALT_T(KC_B));
-
-    set_keymap({mod_tap_key1, mod_tap_key2});
-
-    // Press first mod-tap key.
-    EXPECT_REPORT(driver, (KC_LCTL));
-    mod_tap_key1.press();
-    run_one_scan_loop();
-    VERIFY_AND_CLEAR(driver);
-    EXPECT_EQ(get_speculative_mods(), MOD_BIT_LCTRL);
-
-    // Press second mod-tap key.
-    EXPECT_REPORT(driver, (KC_LCTL, KC_RALT));
-    mod_tap_key2.press();
-    run_one_scan_loop();
-    VERIFY_AND_CLEAR(driver);
-    EXPECT_EQ(get_speculative_mods(), MOD_BIT_LCTRL | MOD_BIT_RALT);
-
-    EXPECT_NO_REPORT(driver);
-    idle_for(TAPPING_TERM + 1);
-    VERIFY_AND_CLEAR(driver);
-    EXPECT_EQ(get_speculative_mods(), 0);
-    EXPECT_EQ(get_mods(), MOD_BIT_LCTRL | MOD_BIT_RALT);
-
-    // Release first mod-tap key.
-    EXPECT_REPORT(driver, (KC_RALT));
-    mod_tap_key1.release();
-    run_one_scan_loop();
-    VERIFY_AND_CLEAR(driver);
-
-    // Release second mod-tap key.
-    EXPECT_EMPTY_REPORT(driver);
-    mod_tap_key2.release();
-    run_one_scan_loop();
-    VERIFY_AND_CLEAR(driver);
-}
-
-TEST_F(SpeculativeHoldDefault, two_mod_taps_same_mods) {
-    TestDriver driver;
-    InSequence s;
-    auto       mod_tap_key1 = KeymapKey(0, 1, 0, GUI_T(KC_A));
-    auto       mod_tap_key2 = KeymapKey(0, 2, 0, GUI_T(KC_B));
-
-    set_keymap({mod_tap_key1, mod_tap_key2});
-
-    // Press first mod-tap key.
-    EXPECT_REPORT(driver, (KC_LGUI));
-    mod_tap_key1.press();
-    run_one_scan_loop();
-    VERIFY_AND_CLEAR(driver);
-
-    // Tap second mod-tap key.
-    EXPECT_NO_REPORT(driver);
-    mod_tap_key2.press();
-    run_one_scan_loop();
-    mod_tap_key2.release();
-    run_one_scan_loop();
-    VERIFY_AND_CLEAR(driver);
-
-    // Release first mod-tap key.
-    EXPECT_REPORT(driver, (KC_LGUI, DUMMY_MOD_NEUTRALIZER_KEYCODE));
-    EXPECT_REPORT(driver, (KC_LGUI));
-    EXPECT_EMPTY_REPORT(driver);
-    EXPECT_REPORT(driver, (KC_A));
-    EXPECT_REPORT(driver, (KC_A, KC_B));
-    EXPECT_REPORT(driver, (KC_A));
-    EXPECT_EMPTY_REPORT(driver);
-    mod_tap_key1.release();
-    run_one_scan_loop();
-    VERIFY_AND_CLEAR(driver);
-}
-
-TEST_F(SpeculativeHoldDefault, respects_get_speculative_hold_callback) {
-    TestDriver driver;
-    InSequence s;
-    auto       mod_tap_key1 = KeymapKey(0, 0, 0, LSFT_T(KC_A));
-    auto       mod_tap_key2 = KeymapKey(0, 1, 0, LSFT_T(KC_B));
-    auto       mod_tap_key3 = KeymapKey(0, 2, 0, LCTL_T(KC_C));
-    auto       mod_tap_key4 = KeymapKey(0, 3, 0, LCTL_T(KC_D));
-    auto       mod_tap_key5 = KeymapKey(0, 4, 0, RSFT_T(KC_E));
-    auto       mod_tap_key6 = KeymapKey(0, 5, 0, RSFT_T(KC_F));
-
-    set_keymap({mod_tap_key1, mod_tap_key2, mod_tap_key3, mod_tap_key4, mod_tap_key5, mod_tap_key6});
-
-    // Enable Speculative Hold selectively for some of the keys.
-    get_speculative_hold_fun = [](uint16_t keycode, keyrecord_t *record) {
-        switch (keycode) {
-            case LSFT_T(KC_B):
-            case LCTL_T(KC_D):
-            case RSFT_T(KC_F):
-                return true;
-        }
-        return false;
-    };
-
-    for (KeymapKey *mod_tap_key : {&mod_tap_key2, &mod_tap_key4, &mod_tap_key6}) {
-        SCOPED_TRACE(std::string("mod_tap_key = ") + mod_tap_key->name);
-        const uint8_t mods = unpack_mod_tap_mods(mod_tap_key->code);
-
-        // Long press and release mod_tap_key.
-        // For these keys where Speculative Hold is enabled, then the mod should
-        // activate immediately on keydown.
-        EXPECT_REPORT(driver, (KC_LCTL + biton(mods)));
-        mod_tap_key->press();
-        run_one_scan_loop();
-        EXPECT_EQ(get_speculative_mods(), mods);
-        EXPECT_EQ(get_mods(), 0);
-        VERIFY_AND_CLEAR(driver);
-
-        EXPECT_NO_REPORT(driver);
-        idle_for(TAPPING_TERM + 1);
-        EXPECT_EQ(get_speculative_mods(), 0);
-        EXPECT_EQ(get_mods(), mods);
-        VERIFY_AND_CLEAR(driver);
-
-        EXPECT_EMPTY_REPORT(driver);
-        mod_tap_key->release();
-        idle_for(TAPPING_TERM + 1);
-        EXPECT_EQ(get_speculative_mods(), 0);
-        EXPECT_EQ(get_mods(), 0);
-        VERIFY_AND_CLEAR(driver);
-    }
-
-    for (KeymapKey *mod_tap_key : {&mod_tap_key1, &mod_tap_key3, &mod_tap_key5}) {
-        SCOPED_TRACE(std::string("mod_tap_key = ") + mod_tap_key->name);
-        const uint8_t mods = unpack_mod_tap_mods(mod_tap_key->code);
-
-        // Long press and release mod_tap_key.
-        // For these keys where Speculative Hold is disabled, the mod should
-        // activate when the key has settled after the tapping term.
-        EXPECT_NO_REPORT(driver);
-        mod_tap_key->press();
-        run_one_scan_loop();
-        EXPECT_EQ(get_speculative_mods(), 0);
-        EXPECT_EQ(get_mods(), 0);
-        VERIFY_AND_CLEAR(driver);
-
-        EXPECT_REPORT(driver, (KC_LCTL + biton(mods)));
-        idle_for(TAPPING_TERM + 1);
-        EXPECT_EQ(get_speculative_mods(), 0);
-        EXPECT_EQ(get_mods(), mods);
-        VERIFY_AND_CLEAR(driver);
-
-        EXPECT_EMPTY_REPORT(driver);
-        mod_tap_key->release();
-        idle_for(TAPPING_TERM + 1);
-        EXPECT_EQ(get_speculative_mods(), 0);
-        EXPECT_EQ(get_mods(), 0);
-        VERIFY_AND_CLEAR(driver);
-    }
-}
-
-TEST_F(SpeculativeHoldDefault, respects_magic_mod_config) {
-    TestDriver driver;
-    InSequence s;
-    auto       mod_tap_key = KeymapKey(0, 1, 0, CTL_T(KC_P));
-
-    set_keymap({mod_tap_key});
-
-    keymap_config.swap_lctl_lgui = true;
-
-    // Press mod-tap-hold key.
-    EXPECT_REPORT(driver, (KC_LGUI));
-    mod_tap_key.press();
-    run_one_scan_loop();
-    VERIFY_AND_CLEAR(driver);
-
-    // Release mod-tap-hold key.
-    EXPECT_REPORT(driver, (KC_LGUI, DUMMY_MOD_NEUTRALIZER_KEYCODE));
-    EXPECT_REPORT(driver, (KC_LGUI));
-    EXPECT_EMPTY_REPORT(driver);
-    EXPECT_REPORT(driver, (KC_P));
-    EXPECT_EMPTY_REPORT(driver);
-    mod_tap_key.release();
-    idle_for(TAPPING_TERM + 1);
-    VERIFY_AND_CLEAR(driver);
-
-    keymap_config.swap_lctl_lgui = false;
-
-    // Press mod-tap-hold key.
-    EXPECT_REPORT(driver, (KC_LCTL));
-    mod_tap_key.press();
-    run_one_scan_loop();
-    VERIFY_AND_CLEAR(driver);
-
-    // Release mod-tap-hold key.
-    EXPECT_EMPTY_REPORT(driver);
-    EXPECT_REPORT(driver, (KC_P));
-    EXPECT_EMPTY_REPORT(driver);
-    mod_tap_key.release();
-    run_one_scan_loop();
     VERIFY_AND_CLEAR(driver);
 }
 
@@ -398,43 +176,6 @@ TEST_F(SpeculativeHoldDefault, tap_regular_key_while_mod_tap_key_is_held) {
 
     // Idle for tapping term of mod tap hold key.
     idle_for(TAPPING_TERM - 3);
-    VERIFY_AND_CLEAR(driver);
-}
-
-TEST_F(SpeculativeHoldDefault, tap_a_mod_tap_key_while_another_mod_tap_key_is_held) {
-    TestDriver driver;
-    InSequence s;
-    auto       first_mod_tap_key  = KeymapKey(0, 1, 0, SFT_T(KC_P));
-    auto       second_mod_tap_key = KeymapKey(0, 2, 0, RSFT_T(KC_A));
-
-    set_keymap({first_mod_tap_key, second_mod_tap_key});
-
-    // Press first mod-tap-hold key.
-    EXPECT_REPORT(driver, (KC_LSFT));
-    first_mod_tap_key.press();
-    run_one_scan_loop();
-    VERIFY_AND_CLEAR(driver);
-
-    // Press second tap-hold key.
-    EXPECT_REPORT(driver, (KC_LSFT, KC_RSFT));
-    second_mod_tap_key.press();
-    run_one_scan_loop();
-    VERIFY_AND_CLEAR(driver);
-
-    // Release second tap-hold key.
-    EXPECT_NO_REPORT(driver);
-    second_mod_tap_key.release();
-    run_one_scan_loop();
-    VERIFY_AND_CLEAR(driver);
-
-    // Release first mod-tap-hold key.
-    EXPECT_EMPTY_REPORT(driver);
-    EXPECT_REPORT(driver, (KC_P));
-    EXPECT_REPORT(driver, (KC_P, KC_A));
-    EXPECT_REPORT(driver, (KC_P));
-    EXPECT_EMPTY_REPORT(driver);
-    first_mod_tap_key.release();
-    run_one_scan_loop();
     VERIFY_AND_CLEAR(driver);
 }
